WritingMiracle/StoryPlot/requirements.txt
Adafruit-GPIO==1.0.3
Adafruit-PCA9685==1.0.1
Adafruit-PureIO==1.1.8
altair==4.1.0
applescript==2020.4.15
appnope==0.1.2
argon2-cffi==20.1.0
asgiref==3.4.1
astor==0.8.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.0
beautifulsoup4==4.8.2
bleach==3.2.1
blinker==1.4
bs4==0.0.1
cachetools==4.2.2
certifi==2019.9.11
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
detect==2019.4.13
Django==3.2.9
djangorestframework==3.12.4
echarts-china-cities-pypkg==0.0.9
entrypoints==0.3
funcy==1.16
future==0.18.2
gensim==4.0.1
gevent==20.6.2
gitdb==4.0.7
GitPython==3.1.18
greenlet==0.4.16
idna==2.8
imageio==2.6.1
importlib-metadata==3.3.0
install==1.3.4
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
ISStreamer==1.0.1
jason==0.1.7
jedi==0.18.0
jieba==0.42.1
Jinja2==2.11.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.7.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.1.0
lml==0.1.0
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.1.3
mistune==0.8.4
MyQR==2.3.1
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
nltk==3.6.2
notebook==6.1.6
numexpr==2.7.3
numpy==1.21.1
only==2019.4.13
opencv-python==4.2.0.32
packaging==20.8
pandas==1.3.1
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pgzero==1.2
pickleshare==0.7.5
Pillow==8.3.1
plotly==5.2.2
prettytable==0.7.2
prometheus-client==0.9.0
prompt-toolkit==3.0.10
protobuf==3.17.3
psutil==5.8.0
ptyprocess==0.7.0
public==2019.4.13
pyarrow==5.0.0
pycparser==2.20
pydeck==0.6.2
pyecharts==1.7.0
pyecharts-jupyter-installer==0.0.3
pygame==1.9.6
Pygments==2.7.3
pyLDAvis==3.3.1
pyparsing==2.4.6
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.3.1
pyzmq==20.0.0
regex==2021.8.3
requests==2.22.0
scikit-learn==0.24.2
scipy==1.7.1
selenium==3.141.0
Send2Trash==1.5.0
simplejson==3.17.0
six==1.14.0
sklearn==0.0
smart-open==5.1.0
smmap==4.0.0
soupsieve==2.0
sqlparse==0.4.2
stop-words==2018.7.23
streamlit==0.87.0
tenacity==8.0.1
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.2.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.62.0
traitlets==5.0.5
typing-extensions==3.7.4.3
tzlocal==3.0
urllib3==1.25.6
validators==0.18.2
wcwidth==0.2.5
webencodings==0.5.1
websocket==0.2.1
widgetsnbextension==3.5.1
wordcloud==1.8.1
zipp==3.4.0
zope.event==4.4
zope.interface==5.1.0
